--- src/CloudCostOptimizer (CCO)/Config_file.json
@@ -1,4 +1,0 @@
-{
-    "Data Extraction (always / onceAday)": "onceAday",
-    "boto3 (enable / disable)": "disable"
-}
