+++ src/CloudCostOptimizer (CCO)/Input_instructions
@@ -1,0 +1,29 @@
+"""
+POST endpoint to get spot fleet hourly price estimations
+body: configuration for fleet, i.e apps,components and other optional configurations
+    dto = {
+        selectedOs: <str>               REQUIRED, requested instances os. options: linux,windows
+        region: <str>                   OPTIONAL, region for instances. options: us-east-2 etc. (see readme for full options)
+        apps: [                         REQUIRED, list of app specifications
+            {
+            app: <str>                 REQUIRED, name of app
+            share: <str>                REQUIRED, set to true if app can share instances with other apps
+            components:[                REQUIRED, list of component specifications
+                {
+                    name: <str>         REQUIRED, name of component
+                    vCPUs: <int>          REQUIRED, required cpu for component
+                    memory: <int>       REQUIRED, required memory for component (GB)
+                    network: <int>      OPTIONAL, component network consumption (GBs)
+                    behavior: <str>     OPTIONAL, component required interruption behavior: options: terminate,stop,hibernation
+                    interruptionFrequency: <int>    OPTIONAL, limit interruption frequency of the instances. options: 0-4 (see readme)
+                    storageSize: <int>  OPTIONAL, component storage size (GB)
+                    IOPS: <int>         OPTIONAL, component required IOPS (MiB I/O)
+                    throughput: <int>   OPTIONAL, component required throughput (MB/s)
+
+                }
+                ]
+            }
+        ]
+
+
+"""
