+++ src/CloudCostOptimizer (CCO)/config_file.json
@@ -1,0 +1,10 @@
+{
+    "Data Extraction (always / onceAday)": "onceAday",
+    "boto3 (enable / disable)": "disable",
+    "Provider (AWS / Azure / Hybrid)": "Hybrid",
+    "Brute Force": "True",
+    "Time per region": 0.21,
+    "Candidate list size": 100,
+    "Proportion amount node/sons": 0.005,
+    "Verbose": "True"
+}
