+++ src/CloudCostOptimizer (CCO)/constants.py
@@ -1,6 +1,6 @@
 """Constants Values- all regions / architecture / physicalProcessor / hardware."""
 
-regions = [
+AWS_REGIONS = [
     "us-east-2",
     "us-east-1",
     "us-west-1",
@@ -23,14 +23,62 @@ regions = [
     "me-south-1",
     "sa-east-1",
 ]
-### regionsSpotPrices = ['us-east-2', 'us-east', 'us-west', 'us-west-2', 'af-south-1','us-west-2-lax-1a',
-###                   'ap-east-1', 'ap-south-1', 'ap-northeast-3', 'ap-northeast-2',
-###                   'apac-sin', 'apac-syd', 'apac-tokyo', 'ca-central-1',
-###                   'eu-central-1', 'eu-ireland', 'eu-west-2', 'eu-south-1',
-###                   'eu-west-3', 'eu-north-1', 'me-south-1', 'sa-east-1']
 
-architecture = ["arm64", "i386", "x86_64", "x86_64_mac"]
-physical_processor = [
+AZURE_REGIONS = [
+    "Central US",
+    "East US",
+    "East US 2",
+    "North Central US",
+    "South Central US",
+    "West Central US",
+    "West US",
+    "West US 2",
+    "West US 3",
+    "UK South",
+    "UK West",
+    "UAE Central",
+    "UAE North",
+    "Switzerland North",
+    "Switzerland West",
+    "Sweden Central",
+    "Sweden South",
+    "Norway East",
+    "Norway West",
+    "Korea Central",
+    "Korea South",
+    "Japan East",
+    "Japan West",
+    "Central India",
+    "South India",
+    "West India",
+    "Germany North",
+    "Germany West Central",
+    "France Central",
+    "France South",
+    "North Europe",
+    "West Europe",
+    "Canada Central",
+    "Canada East",
+    "Brazil South",
+    "Brazil Southeast",
+    "Australia Central",
+    "Australia Central 2",
+    "Australia East",
+    "Australia Southeast",
+    "East Asia",
+    "Southeast Asia",
+    "South Africa North",
+    "South Africa West",
+]
+
+### regionsSpotPrices = ["us-east-2", "us-east", "us-west", "us-west-2", "af-south-1","us-west-2-lax-1a",
+###                   "ap-east-1", "ap-south-1", "ap-northeast-3", "ap-northeast-2",
+###                   "apac-sin", "apac-syd", "apac-tokyo", "ca-central-1",
+###                   "eu-central-1", "eu-ireland", "eu-west-2", "eu-south-1",
+###                   "eu-west-3", "eu-north-1", "me-south-1", "sa-east-1"]
+
+AWS_ARCHITECTURE = ["arm64", "i386", "x86_64", "x86_64_mac"]
+AWS_PHYSICAL_PROCESSOR = [
     "AMD EPYC 7571",
     "Intel Xeon Platinum 8151",
     "Intel Xeon Platinum 8252",
@@ -53,7 +101,8 @@ physical_processor = [
     "Intel Xeon Platinum 8275CL (Cascade Lake)",
     "AWS Graviton2 Processor",
 ]
-hardware = {
+
+AWS_HARDWARE = {
     "gp2": {"IOPS": 250, "throughput": 250, "type": "gp2"},
     "gp3": {"IOPS": 250, "throughput": 1000, "type": "gp3"},
     "piops": {"IOPS": 1000, "throughput": 1000, "type": "piops"},
@@ -62,7 +111,8 @@ hardware = {
     "sc1": {"IOPS": 250, "throughput": 250, "type": "sc1"},
     "previous generation": {"IOPS": 200, "throughput": 90, "type": "magnetic"},
 }
-instances = [
+
+AWS_INSTANCE_TYPES = [
     "m4.2xlarge",
     "g5g.2xlarge",
     "m4.4xlarge",
