+++ src/CloudCostOptimizer (CCO)/group_generator.py
@@ -3,8 +3,26 @@
 from fleet_classes import Offer
 
 
-def partition2(collection, region):
+def generate_all_selections(sets_generator):
+    """Generate all selections."""
+    """given a generator of items that are iterable (set of sets),
+    yield all possible combinations that result from selecting one item from each set.
+    for example: powerset({{a, b}, {c, d}, {e}}) = {{a, c, e}, {a, d, e}, {b, c, e}, {b, d, e}}
+    another example: powerset({}) = {}"""
+    try:
+        first_set = next(sets_generator)
+        # print("got head: ", first_set)
+        for tail_selections in generate_all_selections(sets_generator):
+            # print("got tail: ", tail_selections)
+            for head_selection in first_set:
+                yield head_selection + tail_selections  # maybe [head_selection] + tail_selections ?
+    except StopIteration:
+        yield []
+
+
+def partition2(iterable, region):
     """Partition2 function."""
+    collection = list(iterable)
     if not collection:
         print("Could not find a match in ", region, " region")
     else:
@@ -31,16 +49,24 @@ def partition(collection):  ##first collection- all the group of components
         yield [[first]] + smaller
 
 
-def create_partitions(comp):
-    """Create_partitions function."""
-    return [c for c in partition(comp)]
+# def create_partitions(comp):
+#     """Create_partitions function."""
+#     return [c for c in partition(comp)]
 
 
 def create_groups(
     comp, app_size, region
 ):  ## comp- list of the groups (shared/non-shared), which includes groups' components
     """Create groups function."""
-    partitions = list(
-        map(lambda i: create_partitions(i), comp)
-    )  ## list of all components in each combination
-    return [Offer(p, app_size) for p in partition2(partitions, region)]
+    # original:
+    # partitions = list(
+    #     map(lambda i: partition(i), comp)
+    # )  ## list of all components in each combination
+
+    partitions = map(lambda i: partition(i), comp)
+    # original:
+    # return [Offer(p, app_size) for p in partition2(partitions, region)]
+    # original2:
+    # for p in partition2(partitions, region):
+    for p in generate_all_selections(partitions):
+        yield Offer(p, app_size)
