--- src/CloudCostOptimizer (CCO)/input_Fleet.json
@@ -1,71 +1,0 @@
-{
-    "selectedOs": "linux",
-    "region": "all",
-    "spot/onDemand": "onDemand",
-    "AvailabilityZone": "all",
-    "Architecture": "all",
-    "apps": [
-        {
-            "app": "App1",
-            "share": true,
-            "components": [
-                {
-                    "memory": 8,
-                    "vCPUs": 8,
-                    "network": 5,
-                    "behavior": "terminate",
-                    "frequency": "2",
-                    "storageType": null,
-                    "name": "Comp1"
-                },
-                {
-                    "memory": 8,
-                    "vCPUs": 3,
-                    "network": 0,
-                    "behavior": "hibernate",
-                    "frequency": "3",
-                    "storageType": null,
-                    "burstable": true,
-                    "name": "comp2"
-                },
-                                {
-                    "memory": 8,
-                    "vCPUs": 3,
-                    "network": 0,
-                    "behavior": "hibernate",
-                    "frequency": "3",
-                    "storageType": null,
-                    "burstable": true,
-                    "name": "comp5"
-                }
-            ]
-        },
-        {
-            "app": "App2",
-            "share": false,
-            "components": [
-                {
-                    "memory": 10,
-                    "vCPUs": 5,
-                    "network": 0,
-                    "behavior": "stop",
-                    "frequency": "4",
-                    "storageType": null,
-                    "burstable": true,
-                    "name": "Comp3"
-                }
-            ]
-        },
-        {
-            "app": "App3",
-            "share": false,
-            "components": [
-                {
-                    "memory": 2,
-                    "vCPUs": 2,
-                    "name": "Comp4"
-                }
-            ]
-        }
-    ]
-}
